std: Move change_dir_locked to unstable. #7870
